Complete flex/03 && Align Imgs in flex/04
Finish flex/03 , currently aligned images in flex/04 and will next up
align text in the middle of the div box to reach the look of the image
Readjusting plan will take place incase of not being able to finish this
by the next study session.

Porpose of this commit message is to practice commit messages
structure everynow and then.