Complete flex/05
Make top left close button close shrink 0 to make it not change its look
when page is resized. Same for the Exclamation mark on the top right
of the page.
Make h3 heading for the main message with removing its bottom and top
padding and setting custom ones.
Add padding between text and the bottom buttons
[Practice Commit Messages]