Complete flex/06
Make the body flex coloumn display , with content justified
space-between to achieve the desired look.

Add color to the footer , gave it a color and a flex display
with content justified space-between.

Set content class to flex coloumn with content justified center.
Add padding to buttons to look like the desired output.
[Practice Commit Messages]