Complete flex
Achieve the look of the desired output. Used flex coloumns layout to
set up the header , footer and the midsection.

Used justify-content: space-between; to make make footer at the end of
the page. Header at the start of the page. There was a space between
midsection , header and footer.
Used flex: 1 value to set the midsection to expand to fill the remaining
space.

Removed bullets from unordered lists list-style: none and adjusted the
left padding. Made sure that the anchor element text-decoration is non
and gave it a color.

Board area made it flex: wrap and gave each card a width so depending on
screen size it will react to it

[Practice Commit Messages]